Update common.scad
I happened to be working on the same connectors yesterday night. Alternative method to design the female connector using a chamfer instead of the polygon function. Just some dummy values were used for the female connector since I didn't know the correct values. 

Drawback of the polygon method is that it's cumbersome to design something simple like a chamfer. This method provides more control.

Also a simple method to create a fillet possibly needed later on.